+++ CHANGELOG.md
@@ -10,12 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow setting the `default` tmux color in segment themes. [#296](https://github.com/erikw/tmux-powerline/issues/296).
 ### Fixed
 - The now playing segment is fixed for Last.FM using their 2.0 API. [#307](https://github.com/erikw/tmux-powerline/issues/307)
+- Correctly handle named colours when specifying theme colours. [#314](https://github.com/erikw/tmux-powerline/issues/314)
 
 ## [2.1.0] - 2023-04-16
 ### Added
 - Config options to add keybindigns to mute the status bar added: `TMUX_POWERLINE_MUTE_LEFT_KEYBINDING` and `TMUX_POWERLINE_MUTE_RIGHT_KEYBINDING`.
 ### Changed
-- The old manual way is not supported for simplicity of maintaing the code.
+- The old manual way is not supported for simplicity of maintaining the code.
 
 ## [2.0.0] - 2023-04-15
 ### Added